DORA/adora_chassis_control_dora_node/nodes/odom_viewer/pyproject.toml
[project]
name = "mickrobot-odom-viewer"
version = "0.1.0"
dependencies = [
    "dora-rs==0.3.8",
]

[project.scripts]
odom-viewer-run = "odom_viewer.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"